remove explicit cast

+++ src/WormholeJoin.sol
@@ -193,7 +193,7 @@ contract WormholeJoin {
         bytes32 hashGUID = getGUIDHash(wormholeGUID);
         require(!wormholes[hashGUID].blessed, "WormholeJoin/already-blessed");
         wormholes[hashGUID].blessed = true;
-        wormholes[hashGUID].pending = uint248(wormholeGUID.amount);     // This is an upcast, so it is safe
+        wormholes[hashGUID].pending = wormholeGUID.amount;
         emit Register(hashGUID, wormholeGUID);
         _withdraw(wormholeGUID, hashGUID, maxFee);
     }
